complete Populate DB view post method - add dates to country model

+++ flags/views.py
@@ -1,4 +1,5 @@
-from csv import reader
+from csv import reader, DictWriter
+from datetime import datetime
 
 from django.contrib.auth.mixins import PermissionRequiredMixin
 from django.core.exceptions import ObjectDoesNotExist
@@ -27,14 +28,16 @@ class PopulateDbView(PermissionRequiredMixin, CreateView):
     def get(self, request):
         if not request.user.is_superuser:
             return redirect('index')
-        
-        try:
-            old_pk_list = Country.objects.values_list('id', flat=True)
-        except:
-            old_pk_list = []
 
         try:
-            with open('data.csv') as csv:
+            db = Country.objects.all()
+        except:
+            db = False
+
+        # TODO - revise to also compare CSV data against existing country objects
+
+        try:
+            with open('import_data.csv') as csv:
                 new = 0
                 updated_pk_list = []
                 data_reader = reader(csv)
@@ -46,7 +49,7 @@ class PopulateDbView(PermissionRequiredMixin, CreateView):
                         raise ValueError
                     try:
                         pk_str = row[4]
-                    except:
+                    except IndexError:
                         pk_str = False
                     if pk_str:
                         try:
@@ -60,6 +63,7 @@ class PopulateDbView(PermissionRequiredMixin, CreateView):
             msg = "Error: Review data.csv and fix errors then refresh this page."
             return render(request, self.template_name, { 'message': msg })
         
+        old_pk_list = Country.objects.values_list('id', flat=True).distinct()
         for pk in updated_pk_list:
             if pk not in old_pk_list:
                 msg = f"Error: primary key {pk} does not exist in database."
@@ -78,12 +82,12 @@ class PopulateDbView(PermissionRequiredMixin, CreateView):
             return render(request, self.template_name, { 'message': msg })
         
         msg = f"New entries: {new}<br>Edited entries: {edits}"
-        msg += f"<br>Deleted entries: {len(deletions)}" 
+        msg += f"<br>Deleted entries: {len(deletions)} {old_pk_list}" 
         form = PopulateDbForm()
-        context = { 'form': form, 'message': msg }
+        context = { 'form': form, 'message': msg, 'db': db }
 
-        if deletions:
-            request.session['deletions'] = deletions
+        request.session['deletions'] = deletions if deletions else []
+        
         return render(request, self.template_name, context)
     
     def post(self, request):
@@ -95,7 +99,9 @@ class PopulateDbView(PermissionRequiredMixin, CreateView):
         if not form.is_valid():
             return render(request, self.template_name, { 'form': form })
 
-        with open('data.csv') as csv:
+        # TODO check database before creating new object
+
+        with open('import_data.csv') as csv:
             data_reader = reader(csv)
             next(data_reader)
             for row in data_reader:
@@ -134,14 +140,45 @@ class PopulateDbView(PermissionRequiredMixin, CreateView):
                     )
 
         # Delete items from DB that are not in the updated CSV
-        deletions = request.session.get['deletions']
+        deletions = request.session.get('deletions')
         del request.session['deletions']
 
-        for pk in deletions:
-            country = Country.objects.get(id=pk)
-            country.delete()
+        if deletions:
+            for pk in deletions:
+                country = Country.objects.get(id=pk)
+                country.delete()
         
-        #TODO : create updated CSV of updated DB
+        # Export CSV of updated database
+        export_data = []
+        countries = Country.objects.all()
+        for country in countries:
+            row = {}
+            row['country'] = country.country
+            row['capital'] = country.capital
+            row['country_code'] = country.country_code
+            row['hint'] = country.hint if hasattr(country, 'hint') else ""
+            row['pk'] = country.id
+            export_data.append(row)
+
+        current_time = datetime.now()
+        now = str(current_time)[:17].replace("-", "_").replace(" ", "_").replace(":", "_")
+
+        with open(f"exported_data_{now}.csv", 'w') as export:
+            writer = DictWriter(
+                export,
+                fieldnames=['country', 'capital', 'code', 'hint', 'pk']
+            )
+            writer.writeheader()
+            for row in export_data:
+                writer.writerow(
+                    {
+                        'country': row['country'],
+                        'capital': row['capital'],
+                        'code': row['country_code'],
+                        'hint': row['hint'],
+                        'pk': row['pk']
+                    }
+                )
 
         msg = 'Data successfully posted'
         return render(request, self.template_name, { 'message': msg })
